Docs are updated. Minor refactoring with preprocessors directory names
