+++ tests/all_tests.py
@@ -11,7 +11,7 @@ suite = unittest.TestLoader().loadTestsFromNames(
         'readers.test_reader',  # reading of some directory recursively, concatenates code with '*.c' extension
         'tokenizers.test_tokenizer',  # conversion of source code to tokens string
         'tokenizers.test_c_tokenizer',  # custom logic for conversion of C language
-        'preprocessing.test_c_preprocessing',  # sources modifications for further tokenization
+        'preprocessors.test_c_preprocessing',  # sources modifications for further tokenization
     ]
 )
